modularized verificadora.py code into tools.py and filtercsv.py
